test: split oversized PageArchiver test module by concern
test_page_archiver.py tripped pylint's C0302 too-many-lines and had been silenced with a disable comment; revert that and split it into four cohesive modules (core/config, parallel/ledger, assets/rewrite, archive finalize) instead, each well under the limit. The shared page_archiver fixture moved to conftest.py; a dead unused helper was removed.
